Cleanup, deduplicate layouts
