+++ src/app/tf/[species]/[factor]/histone/route.ts
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import { query } from "../../../../../../lib/client";
-import { DATASETS_QUERY, EPIGENETIC_PROFILE_ACCESSIONS } from "../_ExperimentSelectionPanel/queries";
+import { DATASETS_QUERY, EPIGENETIC_PROFILE_ACCESSIONS } from "../_utility/ExperimentSelectionPanel/queries";
 import { excludeTargetTypes, includeTargetTypes } from "@/consts";
 
 /**
