+++ src/app/tf/[species]/[factor]/layout.tsx
@@ -18,7 +18,7 @@ import Link from "next/link";
 import { ApiContext } from "@/apicontext";
 import { inflate } from "pako";
 import { associateBy } from "queryz";
-import { useParams, usePathname, useRouter } from "next/navigation";
+import { useRouter } from "next/navigation";
 import FactorTabs from "./factortabs";
 import { DeepLearnedSELEXMotifsMetadataQueryResponse } from "./types";
 import { DEEP_LEARNED_MOTIFS_SELEX_METADATA_QUERY } from "./queries";
